refactor: canvas/renderer.js
