+++ sources/state/constants.js
@@ -1,4 +1,5 @@
 // Constants used throughout the application
+export const FRAME_SIZE = 64; // Size of each frame in the spritesheet
 
 // License configuration - single source of truth
 export const LICENSE_CONFIG = [
@@ -58,3 +59,43 @@ export const ANIMATIONS = [
 	{ value: '1h_backslash', label: '1-Handed Backslash' },
 	{ value: '1h_halfslash', label: '1-Handed Halfslash' }
 ];
+
+// Animation offsets (y-positions on spritesheet) - matches chargen.js base_animations
+export const ANIMATION_OFFSETS = {
+  spellcast: 0,
+  thrust: 4 * FRAME_SIZE,
+  walk: 8 * FRAME_SIZE,
+  slash: 12 * FRAME_SIZE,
+  shoot: 16 * FRAME_SIZE,
+  hurt: 20 * FRAME_SIZE,
+  climb: 21 * FRAME_SIZE,
+  idle: 22 * FRAME_SIZE,
+  jump: 26 * FRAME_SIZE,
+  sit: 30 * FRAME_SIZE,
+  emote: 34 * FRAME_SIZE,
+  run: 38 * FRAME_SIZE,
+  combat_idle: 42 * FRAME_SIZE,
+  backslash: 46 * FRAME_SIZE,
+  halfslash: 50 * FRAME_SIZE
+};
+
+// Animation definitions with frame cycles
+export const ANIMATION_CONFIGS = {
+  'spellcast': { row: 0, num: 4, cycle: [0, 1, 2, 3, 4, 5, 6] },
+  'thrust': { row: 4, num: 4, cycle: [0, 1, 2, 3, 4, 5, 6, 7] },
+  'walk': { row: 8, num: 4, cycle: [1, 2, 3, 4, 5, 6, 7, 8] },
+  'slash': { row: 12, num: 4, cycle: [0, 1, 2, 3, 4, 5] },
+  'shoot': { row: 16, num: 4, cycle: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12] },
+  'hurt': { row: 20, num: 1, cycle: [0, 1, 2, 3, 4, 5] },
+  'climb': { row: 21, num: 1, cycle: [0, 1, 2, 3, 4, 5] },
+  'idle': { row: 22, num: 4, cycle: [0, 0, 1] },
+  'jump': { row: 26, num: 4, cycle: [0, 1, 2, 3, 4, 1] },
+  'sit': { row: 30, num: 4, cycle: [0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2] },
+  'emote': { row: 34, num: 4, cycle: [0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2] },
+  'run': { row: 38, num: 4, cycle: [0, 1, 2, 3, 4, 5, 6, 7] },
+  'watering': { row: 4, num: 4, cycle: [0, 1, 4, 4, 4, 4, 5] },
+  'combat': { row: 42, num: 4, cycle: [0, 0, 1] },
+  '1h_slash': { row: 46, num: 4, cycle: [0, 1, 2, 3, 4, 5, 6] },
+  '1h_backslash': { row: 46, num: 4, cycle: [0, 1, 2, 3, 4, 5, 7, 8, 9, 10, 11, 12] },
+  '1h_halfslash': { row: 50, num: 4, cycle: [0, 1, 2, 3, 4, 5] }
+};
